Skip default gcp compute routes.
- These are system generated routes when you create a network. These
will be deleted by the deletion of the network and do not block the
deletion of that network.
- There are no properties on the Route struct that indicate a route is a
default one. Typically, the name will contain the word "default" or the
description will contain the word "Default" but a property like Kind
returns "compute#route" for all routes.
- Why delete routes then? In the case that a user has created a custom
route, they would not be able to delete the network that route belongs
to with leftovers while that route exists.

Fixes #71