Ignoring the path just doesn't ignore it when considering triggers.
- It doesn't include the path at all but we need the ci directory for
task changes so - whatevs, it's getting included. Tried doing the
separate resource (one for leftovers without the ci dir, one for
leftovers that's just the ci dir) and the hh concourse freaked out and
just didn't load the pipeline after a successful set for almost 20
minutes. Going with this instead.